folly/FileUtil.h: fix compiler warning signed/unsigned comparison
Summary:
writeFull() returns ssize_t and without proper casting, comparing
it with data.size() triggers a compiler warning (which is
treated as an error) in the gcc-4.9-glibc-2.20 toolchain.

Test Plan: fbmake runtests

Reviewed By: mhx@fb.com

Subscribers: folly-diffs@, yfeldblum

FB internal diff: D1870710

Signature: t1:1870710:1424874592:f51026c35196d763ad4b192d43c8ccee0255b41d

+++ folly/FileUtil.h
@@ -188,7 +188,7 @@ bool writeFile(const Container& data, const char* filename,
     return false;
   }
   bool ok = data.empty() ||
-    writeFull(fd, &data[0], data.size()) == data.size();
+    writeFull(fd, &data[0], data.size()) == static_cast<ssize_t>(data.size());
   return closeNoInt(fd) == 0 && ok;
 }
